+++ app/src/main/java/com/chs/your_body_profile/domain/model/FoodDetailInfo.kt
@@ -3,7 +3,7 @@ package com.chs.your_body_profile.domain.model
 data class FoodDetailInfo(
     val name: String,
     val servingWeight: Int = 0,
-    val calorie: Int = 0,
+    val calorie: Float = 0f,
     val carbohydrate: Float = 0f,
     val fat: Float = 0f,
     val protein: Float = 0f,
